Feat: error log view 및 API 연동 완료

+++ src/main/java/live/databo3/front/admin/dto/OrganizationDto.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class OrganizationDto {
-    private Long organizationId;
+    private Integer organizationId;
     private String organizationName;
     private String gatewaySn;
     private String controllerSn;
